+++ src/components/QueryEditor/QueryEditor.tsx
@@ -1,0 +1,18 @@
+import React, { PureComponent } from 'react';
+import { QueryEditorProps } from '@grafana/data';
+import { DataSource } from '../../datasource';
+import { DataSourceOptions, Query } from '../../types';
+
+/**
+ * Editor Properties
+ */
+type Props = QueryEditorProps<DataSource, Query, DataSourceOptions>;
+
+/**
+ * Query Editor
+ */
+export class QueryEditor extends PureComponent<Props> {
+  render() {
+    return <div></div>;
+  }
+}