+++ src/module.ts
@@ -1,9 +1,11 @@
 import { DataSourcePlugin } from '@grafana/data';
+import { ConfigEditor, QueryEditor } from './components';
 import { DataSource } from './datasource';
-import { ConfigEditor } from './ConfigEditor';
-import { QueryEditor } from './QueryEditor';
-import { MyQuery, MyDataSourceOptions } from './types';
+import { DataSourceOptions, Query } from './types';
 
-export const plugin = new DataSourcePlugin<DataSource, MyQuery, MyDataSourceOptions>(DataSource)
+/**
+ * Datasource Plugin
+ */
+export const plugin = new DataSourcePlugin<DataSource, Query, DataSourceOptions>(DataSource)
   .setConfigEditor(ConfigEditor)
   .setQueryEditor(QueryEditor);