+++ src/types.ts
@@ -1,26 +1,18 @@
 import { DataQuery, DataSourceJsonData } from '@grafana/data';
 
-export interface MyQuery extends DataQuery {
-  queryText?: string;
-  constant: number;
-  withStreaming: boolean;
-}
-
-export const defaultQuery: Partial<MyQuery> = {
-  constant: 6.5,
-  withStreaming: false,
-};
+/**
+ * Query
+ */
+export interface Query extends DataQuery {}
 
 /**
  * These are options configured for each DataSource instance.
  */
-export interface MyDataSourceOptions extends DataSourceJsonData {
-  path?: string;
-}
-
-/**
- * Value that is used in the backend, but never sent over HTTP to the frontend
- */
-export interface MySecureJsonData {
-  apiKey?: string;
+export interface DataSourceOptions extends DataSourceJsonData {
+  /**
+   * Filter
+   *
+   * @type {string}
+   */
+  filter?: string;
 }