+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
-## 1.0.0 (Unreleased)
+## 1.0.0 (IN PROGRESS)
 
-Initial release.
+- Initial release based on the Grafana 8.4.1.
+- Returns Environment Variables.
+- Allows to filter unnecessary or secured variables using Regex.